+++ frontend/src/lib/components/apps/components/layout/AppList.svelte
@@ -204,7 +204,7 @@
 						page = page + 1
 						outputs?.page.set(page, true)
 					}}
-					disabled={pagination.disableNext}
+					disabled={pagination.disableNext && pagination.total > 0}
 				>
 					Next
 
@@ -214,7 +214,7 @@
 						<ChevronRight size={14} />
 					{/if}
 				</Button>
-				<div class="text-xs">{page + 1} of {pagination.total}</div>
+				<div class="text-xs">{page + 1} {pagination.total > 0 ? `of ${pagination.total}` : ''}</div>
 			</div>
 		{/if}
 	</div>